+++ nanairo-reader-app/src/main/java/eu/nanairo_reader/data/dao/ItemDaoMock.java
@@ -5,13 +5,13 @@ import java.util.List;
 
 import eu.nanairo_reader.data.entity.ItemEntity;
 
-public class ItemDaoMock extends BaseDaoMock<ItemEntity, Integer> implements ItemDao {
+public class ItemDaoMock extends BaseDaoMock<ItemEntity, Long> implements ItemDao {
 	@Override
-	public List<ItemEntity> getList(int id) {
+	public List<ItemEntity> getList(long id) {
 		List<ItemEntity> list = new ArrayList<ItemEntity>();
 
 		ItemEntity item1 = new ItemEntity();
-		item1.setId(1);
+		item1.setId(1L);
 		item1.setTitle("6月16日(日)");
 		item1.setLink("http://tbn17.com/archives/1764915.html");
 		item1.setContent("bbb<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aaaa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa<br>aa");
@@ -19,7 +19,7 @@ public class ItemDaoMock extends BaseDaoMock<ItemEntity, Integer> implements Ite
 		list.add(item1);
 
 		ItemEntity item2 = new ItemEntity();
-		item2.setId(2);
+		item2.setId(2L);
 		item2.setTitle("Google");
 		item2.setLink("http://tbn17.com/archives/1764915.html");
 		item2.setContent("bbb<br>aa");
@@ -28,13 +28,13 @@ public class ItemDaoMock extends BaseDaoMock<ItemEntity, Integer> implements Ite
 
 		if (1 == id) {
 			ItemEntity item3 = new ItemEntity();
-			item3.setId(3);
+			item3.setId(3L);
 			item3.setTitle("y");
 			item3.setMidoku(0);
 			list.add(item3);
 		} else {
 			ItemEntity item3 = new ItemEntity();
-			item3.setId(3);
+			item3.setId(3L);
 			item3.setTitle("g");
 			item3.setMidoku(0);
 			list.add(item3);
@@ -44,7 +44,7 @@ public class ItemDaoMock extends BaseDaoMock<ItemEntity, Integer> implements Ite
 	}
 
 	@Override
-	public int getMidokuCount(Integer id) {
+	public int getMidokuCount(long id) {
 		return 3;
 	}
 }
