+++ nanairo-reader-app/src/main/java/eu/nanairo_reader/data/dao/SubscriptionDao.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 import eu.nanairo_reader.data.entity.SubscriptionEntity;
 
-public interface SubscriptionDao extends BaseDao<SubscriptionEntity, Integer> {
+public interface SubscriptionDao extends BaseDao<SubscriptionEntity, Long> {
 	List<SubscriptionEntity> getList();
 }
