+++ nanairo-reader-app/src/main/java/eu/nanairo_reader/data/dao/SubscriptionDaoImpl.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import eu.nanairo_reader.data.entity.SubscriptionEntity;
 
-public class SubscriptionDaoImpl extends BaseDaoImpl<SubscriptionEntity, Integer> implements SubscriptionDao {
+public class SubscriptionDaoImpl extends BaseDaoImpl<SubscriptionEntity, Long> implements SubscriptionDao {
 	protected Class<SubscriptionEntity> getEntityClass() {
 		return SubscriptionEntity.class;
 	}
